Add StaticCompositeResolver
